fix : GlobalPos 오류 수정 (#19)
* fix : globalPos 자료형 변경

int형에서 float형으로 변경

* feat : image클래스의 Render함수 변경

화면에 그리는 좌표를 int값이 아닌 float값으로 받도록 변경

* fix : globalPos가 바뀔 때 움직이는 오브젝트들의 움직임 수정

움직일 때 globalPos를 int형으로 변환하는 부분 삭제

+++ FireFlower.cpp
@@ -93,6 +93,6 @@ void FireFlower::AutoMove()
 {
     if (PLAYER->GetPos().x + PLAYER->GetCurrSpeed() > WIN_SIZE_X / 2)
     {
-        pos.x -= (int)PLAYER->GetCurrSpeed();
+        pos.x -= PLAYER->GetCurrSpeed();
     }
 }
