+++ Gunba.cpp
@@ -78,7 +78,7 @@ void Gunba::UpdatePosition()
 {
     if (PLAYER->GetCurrSpeed() + PLAYER->GetPos().x > WIN_SIZE_X / 2)
     {
-        pos.x += speed * DELETA_TIME - (int)PLAYER->GetCurrSpeed();
+        pos.x += speed * DELETA_TIME - PLAYER->GetCurrSpeed();
     }
     else
     {
